Combine types
