+++ shared/src/types/index.ts
@@ -2,4 +2,3 @@ export * from "./actions";
 export * from "./types";
 export * from "./auth";
 export * from "./messages";
-export * from "./messageTypes";
