+++ shared/src/types/messages.ts
@@ -12,6 +12,28 @@ import {
   HubConfig,
 } from "./types";
 
+export const MessageTypes = {
+  FULL_STATE_UPDATE: "FULL_STATE_UPDATE",
+
+  ANALYSIS_STATE_UPDATE: "ANALYSIS_STATE_UPDATE",
+
+  CHAT_MESSAGES_UPDATE: "CHAT_MESSAGES_UPDATE",
+  CHAT_MESSAGE_STREAMING_UPDATE: "CHAT_MESSAGE_STREAMING_UPDATE",
+  CHAT_STREAMING_CHUNK: "CHAT_STREAMING_CHUNK",
+
+  PROFILES_UPDATE: "PROFILES_UPDATE",
+
+  SERVER_STATE_UPDATE: "SERVER_STATE_UPDATE",
+
+  SOLUTION_WORKFLOW_UPDATE: "SOLUTION_WORKFLOW_UPDATE",
+
+  CONFIG_ERRORS_UPDATE: "CONFIG_ERRORS_UPDATE",
+  DECORATORS_UPDATE: "DECORATORS_UPDATE",
+  SETTINGS_UPDATE: "SETTINGS_UPDATE",
+} as const;
+
+export type MessageType = (typeof MessageTypes)[keyof typeof MessageTypes];
+
 /**
  * Granular message types for VSCode extension -> Webview communication
  *
